requirements.txt
black
flake8
mypy
pylint

gym==0.13
jupyter
numpy
matplotlib
metaflow
pyyaml
tensorflow==2.1
git+git://github.com/openai/baselines@tf2#egg